+++ lib/pages/start_page/card_start_game.dart
